Convert most components, refactor gamepad displays
