+++ package.json
@@ -57,6 +57,7 @@
     ]
   },
   "devDependencies": {
+    "@types/jmuxer": "^2.0.7",
     "@types/node": "^25.3.3",
     "@types/piexifjs": "^1.0.0",
     "@types/react-gamepad": "^1.0.6"
