+++ src/components/input/KeyboardController.tsx
@@ -1,9 +1,10 @@
 import {useEffect} from 'react'
 import {useDispatch} from 'react-redux'
 import {keyPressed, keyReleased} from '../../store/inputSlice.js'
+import type { RoverDispatch } from '../../store/store.js'
 
 export const KeyboardController = () => {
-  const dispatch = useDispatch()
+  const dispatch = useDispatch<RoverDispatch>()
 
   useEffect(() => {
     const handleKeyDown = (event: KeyboardEvent) => dispatch(keyPressed({key: event.key}))
