--- src/components/navigation/ScienceStatus.jsx
@@ -1,13 +1,0 @@
-import {useSelector} from 'react-redux'
-import {getSpeed} from '../../store/inputSlice.js'
-
-function ScienceStatus() {
-  const speed = useSelector(getSpeed)
-  return (
-    <p>
-      Current linkage speed: <br /> {speed}%
-    </p>
-  )
-}
-
-export default ScienceStatus
