--- src/components/roverModel/RoverModel.css
@@ -1,3 +1,0 @@
-.arm-model {
-  background-color: #222;
-}
