--- src/components/telemetry/InputTelemetry.css
@@ -1,3 +1,0 @@
-.input-telemetry {
-  flex: 0.5;
-}
