--- src/components/telemetry/TelemetryPanel.jsx
@@ -1,14 +1,0 @@
-import JointTelemetry from './JointTelemetry'
-// import InputTelemetry from './InputTelemetry'
-import './TelemetryPanel.css'
-
-function TelemetryPanel() {
-  return (
-    <div className="telemetry-panel">
-      {/* <InputTelemetry /> */}
-      <JointTelemetry />
-    </div>
-  )
-}
-
-export default TelemetryPanel
