+++ src/components/telemetry/TelemetryPanel.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+import {JointTelemetry} from './JointTelemetry.js'
+import './TelemetryPanel.css'
+
+export const TelemetryPanel = () => {
+  return (
+    <div className="telemetry-panel">
+      <JointTelemetry />
+    </div>
+  )
+}
