--- src/types/legacy-js-components.d.ts
@@ -1,6 +1,0 @@
-declare module '*.jsx' {
-  import type {ComponentType} from 'react'
-
-  const Component: ComponentType<any>
-  export default Component
-}
